+++ tests/test_vjp.py
@@ -229,6 +229,7 @@ class TestVectorJacobianProduct:
                 dev.vjp(tape.measurements, dy)(tape)
 
     @pytest.mark.skipif(ld._new_API, reason="Old API required")
+    @pytest.mark.usefixtures("use_legacy_and_new_opmath")
     def test_proj_unsupported(self, dev):
         """Test if a QuantumFunctionError is raised for a Projector observable"""
 
@@ -246,7 +247,7 @@ class TestVectorJacobianProduct:
 
         with qml.tape.QuantumTape() as tape:
             qml.CRX(0.1, wires=[0, 1])
-            qml.expval(qml.Projector([0], wires=[0]) @ qml.PauliZ(0))
+            qml.expval(qml.Projector([0], wires=[0]) @ qml.PauliZ(1))
 
         with pytest.raises(
             qml.QuantumFunctionError,