+++ .github/CHANGELOG.md
@@ -27,6 +27,9 @@
 * Add Vector-Jacobian Product calculation support to `lightning.qubit`.
   [(#644)](https://github.com/PennyLaneAI/pennylane-lightning/pull/644)
 
+* Add support for using new operator arithmetic as the default.
+  [(#649)](https://github.com/PennyLaneAI/pennylane-lightning/pull/649)
+
 ### Breaking changes
 
 * Migrate `lightning.qubit` to the new device API.
@@ -52,6 +55,9 @@
 
 * Fix random `coverage xml` CI issues.
   [(#635)](https://github.com/PennyLaneAI/pennylane-lightning/pull/635)
+
+* `lightning.qubit` correctly decomposed state preparation operations with adjoint differentiation.
+  [(#661)](https://github.com/PennyLaneAI/pennylane-lightning/pull/661)
 
 ### Contributors